requirements.txt
streamlit
pandas
numpy
opencv-python-headless
pytesseract
Pillow
XlsxWriter